Read/Write config.json
Able to read and write multiple pages of options to config files. Upon creation, the preview images will be copied and appropriately named. Until an option name is entered on a page, everything except the back button and cancel button will be disabled. This is in order to prevent blank pages from being stored and non-blank pages from being discarded.